@@ -1,6 +1,5 @@
 {
   "name": "hls.js",
-  "version": "0.10.1",
   "license": "Apache-2.0",
   "description": "JavaScript HLS client using MediaSourceExtension",
   "homepage": "https://github.com/video-dev/hls.js",
@@ -26,56 +25,45 @@
   },
   "scripts": {
     "build": "webpack --progress",
-    "build:watch": "webpack --progress --watch",
+    "build:debug": "webpack --progress --env.debug",
+    "build:watch": "webpack --progress --env.debug --watch",
     "build:analyze": "ANALYZE=true webpack --progress",
-    "build:release": "npm run build && npm run test && git add dist/* && git commit -m 'Update dist' && npm run docs:release",
-    "commit:release": "npm run build:release && git add dist/* && git commit -m 'update dist'",
     "dev": "webpack-dev-server --progress --env.debug --port 8000",
     "docs": "npm run docs:generate",
     "docs:generate": "esdoc",
     "docs:clean": "rm -Rf html/docs",
     "docs:update": "git add docs/html && git commit docs/html -m 'Update docs'",
     "docs:release": "npm run docs:clean && npm run docs:generate && npm run docs:update",
-    "lint": "npm run lint:src && npm run lint:tests",
-    "lint:fix": "eslint src/ tests/ --fix",
-    "lint:quiet": "eslint src/ tests/ --quiet",
-    "lint:src": "eslint src/",
-    "lint:tests": "eslint tests/",
-    "lint:demo": "eslint demo/",
-    "precommit": "./scripts/precommit",
+    "lint": "./scripts/lint.sh",
+    "lint:fix": "./scripts/lint.sh --fix",
+    "lint:quiet": "./scripts/lint.sh --quiet",
+    "precommit": "./scripts/precommit.sh",
     "pretest": "npm run lint",
-    "release:pre": "mversion prerelease && npm run build:release",
-    "release:patch": "mversion p && npm run build:release",
-    "release:minor": "mversion mi && npm run build:release",
-    "release:major": "mversion ma && npm run build:release",
     "start": "npm run dev",
     "test": "npm run test:unit && npm run test:func",
     "test:unit": "karma start karma.conf.js",
     "test:unit:watch": "karma start karma.conf.js --auto-watch --no-single-run",
-    "test:func": "cross-env BABEL_ENV=test mocha --compilers js:babel-register tests/functional/auto/setup.js --timeout 40000"
+    "test:func": "mocha --require ts-node/register tests/functional/auto/setup.js --timeout 40000"
   },
   "dependencies": {
+    "events": "^3.0.0",
     "string.prototype.endswith": "^0.2.0",
     "url-toolkit": "^2.1.2"
   },
   "devDependencies": {
     "arraybuffer-equal": "^1.0.4",
-    "babel-core": "^6.25.0",
-    "babel-loader": "^7.1.1",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-register": "^6.24.0",
     "chromedriver": "^2.38.3",
-    "cross-env": "^5.0.2",
     "deep-strict-equal": "^0.2.0",
     "esdoc": "^1.1.0",
     "esdoc-standard-plugin": "^1.0.0",
+    "esdoc-typescript-plugin": "^1.0.1",
     "eslint": "^4.13.1",
     "eslint-config-standard": "^11.0.0",
     "eslint-plugin-import": "^2.9.0",
     "eslint-plugin-node": "^6.0.1",
     "eslint-plugin-promise": "^3.6.0",
     "eslint-plugin-standard": "^3.0.1",
+    "eslint-plugin-typescript": "^0.12.0",
     "http-server": "^0.11.0",
     "husky": "^0.14.3",
     "istanbul-instrumenter-loader": "^3.0.1",
@@ -88,15 +76,21 @@
     "karma-should": "^1.0.0",
     "karma-sinon": "^1.0.5",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^2.0.4",
+    "karma-webpack": "^3.0.0",
     "mocha": "^3.0.2",
     "mversion": "^1.10.1",
+    "npm-run-all": "^4.1.3",
     "selenium-webdriver": "^3.1.0",
     "should": "^13.1.3",
     "sinon": "^4.1.3",
-    "webpack": "^3.5.5",
+    "ts-loader": "^4.4.2",
+    "ts-node": "^6.1.0",
+    "typescript": "^2.9.1",
+    "typescript-eslint-parser": "^16.0.0",
+    "webpack": "^4.11.1",
     "webpack-bundle-analyzer": "^2.9.1",
-    "webpack-dev-server": "^2.7.1",
-    "webworkify-webpack": "^2.0.0"
+    "webpack-cli": "^3.0.2",
+    "webpack-dev-server": "^3.1.4",
+    "webworkify-webpack": "^2.1.2"
   }
 }
