@@ -40,7 +40,14 @@ class AudioTrackController extends TaskLoop {
      * Currently selected index in `tracks`
      * @member {number} trackId
      */
-    this.trackId = -1;
+    this._trackId = -1;
+
+    /**
+     * @private
+     * If should select tracks according to default track attribute
+     * @member {boolean} _selectDefaultTrack
+     */
+    this._selectDefaultTrack = true;
 
     /**
      * @public
@@ -70,7 +77,8 @@ class AudioTrackController extends TaskLoop {
    */
   onManifestLoading () {
     this.tracks = [];
-    this.trackId = -1;
+    this._trackId = -1;
+    this._selectDefaultTrack = true;
   }
 
   /**
@@ -195,15 +203,25 @@ class AudioTrackController extends TaskLoop {
    * @type {number} Index into audio-tracks list of currently selected track.
    */
   get audioTrack () {
-    return this.trackId;
+    return this._trackId;
   }
 
   /**
    * Select current track by index
    */
   set audioTrack (newId) {
+    this._setAudioTrack(newId);
+    // If audio track is selected from API then don't choose from the manifest default track
+    this._selectDefaultTrack = false;
+  }
+
+  /**
+   * @private
+   * @param {number} newId
+   */
+  _setAudioTrack (newId) {
     // noop on same audio track id as already set
-    if (this.trackId === newId && this.tracks[this.trackId].details) {
+    if (this._trackId === newId && this.tracks[this._trackId].details) {
       logger.debug('Same id as current audio-track passed, and track details available -> no-op');
       return;
     }
@@ -220,7 +238,7 @@ class AudioTrackController extends TaskLoop {
 
     // stopping live reloading timer if any
     this.clearInterval();
-    this.trackId = newId;
+    this._trackId = newId;
 
     const { url, type, id } = audioTrack;
     this.hls.trigger(Event.AUDIO_TRACK_SWITCHING, { id, type, url });
@@ -231,7 +249,7 @@ class AudioTrackController extends TaskLoop {
    * @override
    */
   doTick () {
-    this._updateTrack(this.trackId);
+    this._updateTrack(this._trackId);
   }
 
   /**
@@ -244,7 +262,7 @@ class AudioTrackController extends TaskLoop {
       return;
     }
 
-    const currentAudioTrack = this.tracks[this.trackId];
+    const currentAudioTrack = this.tracks[this._trackId];
 
     let name = null;
     if (currentAudioTrack) {
@@ -252,11 +270,13 @@ class AudioTrackController extends TaskLoop {
     }
 
     // Pre-select default tracks if there are any
-    const defaultTracks = tracks.filter((track) => track.default);
-    if (defaultTracks.length) {
-      tracks = defaultTracks;
-    } else {
-      logger.warn('No default audio tracks defined');
+    if (this._selectDefaultTrack) {
+      const defaultTracks = tracks.filter((track) => track.default);
+      if (defaultTracks.length) {
+        tracks = defaultTracks;
+      } else {
+        logger.warn('No default audio tracks defined');
+      }
     }
 
     let trackFound = false;
@@ -269,11 +289,11 @@ class AudioTrackController extends TaskLoop {
         }
         // We need to match the (pre-)selected group ID
         // and the NAME of the current track.
-        if ((!this.audioGroupId || track.groupId === this.audioGroupId) &&
+        if ((!this.audioGroupId || track.groupId === this.audioGroupId) &&
           (!name || name === track.name)) {
           // If there was a previous track try to stay with the same `NAME`.
           // It should be unique across tracks of same group, and consistent through redundant track groups.
-          this.audioTrack = track.id;
+          this._setAudioTrack(track.id);
           trackFound = true;
         }
       });
@@ -337,7 +357,7 @@ class AudioTrackController extends TaskLoop {
 
     // stopping live reloading timer if any
     this.clearInterval();
-    this.trackId = newId;
+    this._trackId = newId;
     logger.log(`trying to update audio-track ${newId}`);
     const audioTrack = this.tracks[newId];
     this._loadTrackDetailsIfNeeded(audioTrack);
@@ -348,10 +368,10 @@ class AudioTrackController extends TaskLoop {
    */
   _handleLoadError () {
     // First, let's black list current track id
-    this.trackIdBlacklist[this.trackId] = true;
+    this.trackIdBlacklist[this._trackId] = true;
 
     // Let's try to fall back on a functional audio-track with the same group ID
-    const previousId = this.trackId;
+    const previousId = this._trackId;
     const { name, language, groupId } = this.tracks[previousId];
 
     logger.warn(`Loading failed on audio track id: ${previousId}, group-id: ${groupId}, name/language: "${name}" / "${language}"`);
@@ -377,7 +397,7 @@ class AudioTrackController extends TaskLoop {
 
     logger.log('Attempting audio-track fallback id:', newId, 'group-id:', this.tracks[newId].groupId);
 
-    this.audioTrack = newId;
+    this._setAudioTrack(newId);
   }
 }
 
