@@ -16,10 +16,19 @@ import MP4Remuxer from '../remux/mp4-remuxer';
 import PassThroughRemuxer from '../remux/passthrough-remuxer';
 
 import { getSelfScope } from '../utils/get-self-scope';
+import { logger } from '../utils/logger';
 
 // see https://stackoverflow.com/a/11237259/589493
 const global = getSelfScope(); // safeguard for code that might run both on worker and main thread
-const performance = global;
+
+let now;
+// performance.now() not available on WebWorker, at least on Safari Desktop
+try {
+  now = global.performance.now.bind(global.performance);
+} catch (err) {
+  logger.debug('Unable to use Performance API on this environment');
+  now = global.Date.now;
+}
 
 class DemuxerInline {
   constructor (observer, typeSupported, config, vendor) {
@@ -43,23 +52,11 @@ class DemuxerInline {
         decrypter = this.decrypter = new Decrypter(this.observer, this.config);
       }
 
-      let localthis = this;
-      // performance.now() not available on WebWorker, at least on Safari Desktop
-      let startTime;
-      try {
-        startTime = performance.now();
-      } catch (error) {
-        startTime = Date.now();
-      }
-      decrypter.decrypt(data, decryptdata.key.buffer, decryptdata.iv.buffer, function (decryptedData) {
-        let endTime;
-        try {
-          endTime = performance.now();
-        } catch (error) {
-          endTime = Date.now();
-        }
-        localthis.observer.trigger(Event.FRAG_DECRYPTED, { stats: { tstart: startTime, tdecrypt: endTime } });
-        localthis.pushDecrypted(new Uint8Array(decryptedData), decryptdata, new Uint8Array(initSegment), audioCodec, videoCodec, timeOffset, discontinuity, trackSwitch, contiguous, duration, accurateTimeOffset, defaultInitPTS);
+      const startTime = now();
+      decrypter.decrypt(data, decryptdata.key.buffer, decryptdata.iv.buffer, (decryptedData) => {
+        const endTime = now();
+        this.observer.trigger(Event.FRAG_DECRYPTED, { stats: { tstart: startTime, tdecrypt: endTime } });
+        this.pushDecrypted(new Uint8Array(decryptedData), decryptdata, new Uint8Array(initSegment), audioCodec, videoCodec, timeOffset, discontinuity, trackSwitch, contiguous, duration, accurateTimeOffset, defaultInitPTS);
       });
     } else {
       this.pushDecrypted(new Uint8Array(data), decryptdata, new Uint8Array(initSegment), audioCodec, videoCodec, timeOffset, discontinuity, trackSwitch, contiguous, duration, accurateTimeOffset, defaultInitPTS);
