@@ -6,7 +6,8 @@
 import DemuxerInline from '../demux/demuxer-inline';
 import Event from '../events';
 import { enableLogs } from '../utils/logger';
-import EventEmitter from 'events';
+
+import { EventEmitter } from 'events';
 
 let DemuxerWorker = function (self) {
   // observer setup
@@ -28,13 +29,11 @@ let DemuxerWorker = function (self) {
     // console.log('demuxer cmd:' + data.cmd);
     switch (data.cmd) {
     case 'init':
-      let config = JSON.parse(data.config);
+      const config = JSON.parse(data.config);
       self.demuxer = new DemuxerInline(observer, data.typeSupported, config, data.vendor);
-      try {
-        enableLogs(config.debug === true);
-      } catch (err) {
-        console.warn('demuxerWorker: unable to enable logs');
-      }
+
+      enableLogs(config.debug);
+
       // signal end of worker init
       forwardMessage('init', null);
       break;
