@@ -68,10 +68,33 @@ export function adjustPts (sliding, details) {
   details.PTSKnown = true;
 }
 
-// If a change in CC is detected, the PTS can no longer be relied upon
-// Attempt to align the level by using the last level - find the last frag matching the current CC and use it's PTS
-// as a reference
-export function alignDiscontinuities (lastFrag, lastLevel, details) {
+/**
+ * Using the parameters of the last level, this function computes PTS' of the new fragments so that they form a
+ * contiguous stream with the last fragments.
+ * The PTS of a fragment lets Hls.js know where it fits into a stream - by knowing every PTS, we know which fragment to
+ * download at any given time. PTS is normally computed when the fragment is demuxed, so taking this step saves us time
+ * and an extra download.
+ * @param lastFrag
+ * @param lastLevel
+ * @param details
+ */
+export function alignStream (lastFrag, lastLevel, details) {
+  alignDiscontinuities(lastFrag, details, lastLevel);
+  if (!details.PTSKnown && lastLevel) {
+    // If the PTS wasn't figured out via discontinuity sequence that means there was no CC increase within the level.
+    // Aligning via Program Date Time should therefore be reliable, since PDT should be the same within the same
+    // discontinuity sequence.
+    alignPDT(details, lastLevel.details);
+  }
+}
+
+/**
+ * Computes the PTS if a new level's fragments using the PTS of a fragment in the last level which shares the same
+ * discontinuity sequence.
+ * @param lastLevel - The details of the last loaded level
+ * @param details - The details of the new level
+ */
+export function alignDiscontinuities (lastFrag, details, lastLevel) {
   if (shouldAlignOnDiscontinuities(lastFrag, lastLevel, details)) {
     const referenceFrag = findDiscontinuousReferenceFrag(lastLevel.details, details);
     if (referenceFrag) {
@@ -79,16 +102,26 @@ export function alignDiscontinuities (lastFrag, lastLevel, details) {
       adjustPts(referenceFrag.start, details);
     }
   }
-  // try to align using programDateTime attribute (if available)
-  if (details.PTSKnown === false && lastLevel && lastLevel.details && lastLevel.details.fragments && lastLevel.details.fragments.length) {
+}
+
+/**
+ * Computes the PTS of a new level's fragments using the difference in Program Date Time from the last level.
+ * @param details - The details of the new level
+ * @param lastDetails - The details of the last loaded level
+ */
+export function alignPDT (details, lastDetails) {
+  if (lastDetails && lastDetails.fragments.length) {
+    if (!details.hasProgramDateTime || !lastDetails.hasProgramDateTime) {
+      return;
+    }
     // if last level sliding is 1000 and its first frag PROGRAM-DATE-TIME is 2017-08-20 1:10:00 AM
     // and if new details first frag PROGRAM DATE-TIME is 2017-08-20 1:10:08 AM
     // then we can deduce that playlist B sliding is 1000+8 = 1008s
-    let lastPDT = lastLevel.details.programDateTime;
-    let newPDT = details.programDateTime;
+    let lastPDT = lastDetails.fragments[0].programDateTime;
+    let newPDT = details.fragments[0].programDateTime;
     // date diff is in ms. frag.start is in seconds
-    let sliding = (newPDT - lastPDT) / 1000 + lastLevel.details.fragments[0].start;
-    if (!isNaN(sliding)) {
+    let sliding = (newPDT - lastPDT) / 1000 + lastDetails.fragments[0].start;
+    if (Number.isFinite(sliding)) {
       logger.log(`adjusting PTS using programDateTime delta, sliding:${sliding.toFixed(3)}`);
       adjustPts(sliding, details);
     }
