@@ -1,9 +1,9 @@
-const assert = require('assert');
-
 import Hls from '../../../src/hls';
 import Event from '../../../src/events';
 import { FragmentTracker, FragmentState } from '../../../src/controller/fragment-tracker';
 import PlaylistLoader from '../../../src/loader/playlist-loader';
+
+const assert = require('assert');
 
 const LevelType = PlaylistLoader.LevelType;
 
