@@ -16,6 +16,7 @@ describe('StreamController tests', function () {
     hls = new Hls({});
     fragmentTracker = new FragmentTracker(hls);
     streamController = new StreamController(hls, fragmentTracker);
+    streamController.startFragRequested = true;
   });
 
   /**
@@ -77,11 +78,10 @@ describe('StreamController tests', function () {
     });
   });
 
-  describe('PDT vs SN tests for discontinuities with PDT', function () {
-    let PDT = 'Fri Sep 15 2017 12:11:01:523 GMT-0700 (Pacific Daylight Time)';
+  describe('SN Searching', function () {
     let fragPrevious = {
-      pdt: 1505502671523,
-      endPdt: 1505502676523,
+      programDateTime: 1505502671523,
+      endProgramDateTime: 1505502676523,
       duration: 5.000,
       level: 1,
       start: 10.000,
@@ -92,64 +92,54 @@ describe('StreamController tests', function () {
     let fragLen = mockFragments.length;
     let levelDetails = {
       startSN: mockFragments[0].sn,
-      endSN: mockFragments[mockFragments.length - 1].sn,
-      programDateTime: null // If this field is null SN search is used by default, if set is PDT
+      endSN: mockFragments[mockFragments.length - 1].sn
     };
     let bufferEnd = fragPrevious.start + fragPrevious.duration;
     let end = mockFragments[mockFragments.length - 1].start + mockFragments[mockFragments.length - 1].duration;
 
-    it('SN search choosing wrong fragment (3 instead of 2) after level loaded', function () {
-      levelDetails.programDateTime = null;
+    before(function () {
+      levelDetails.hasProgramDateTime = false;
+    });
 
+    it('PTS search choosing wrong fragment (3 instead of 2) after level loaded', function () {
       let foundFragment = streamController._findFragment(0, fragPrevious, fragLen, mockFragments, bufferEnd, end, levelDetails);
       let resultSN = foundFragment ? foundFragment.sn : -1;
       assert.equal(foundFragment, mockFragments[3], 'Expected sn 3, found sn segment ' + resultSN);
     });
 
     // TODO: This test fails if using a real instance of Hls
-    it('SN search choosing the right segment if fragPrevious is not available', function () {
-      levelDetails.programDateTime = null;
-
+    it('PTS search choosing the right segment if fragPrevious is not available', function () {
       let foundFragment = streamController._findFragment(0, null, fragLen, mockFragments, bufferEnd, end, levelDetails);
       let resultSN = foundFragment ? foundFragment.sn : -1;
       assert.equal(foundFragment, mockFragments[3], 'Expected sn 2, found sn segment ' + resultSN);
     });
 
-    it('PDT search choosing fragment after level loaded', function () {
-      levelDetails.programDateTime = PDT;// If programDateTime contains a date then PDT is used
-
-      let foundFragment = streamController._findFragment(0, fragPrevious, fragLen, mockFragments, bufferEnd, end, levelDetails);
-      let resultSN = foundFragment ? foundFragment.sn : -1;
-      assert.equal(foundFragment, mockFragments[3], 'Expected sn 3, found sn segment ' + resultSN);
+    it('returns the last fragment if the stream is fully buffered', function () {
+      const actual = streamController._findFragment(0, null, mockFragments.length, mockFragments, end, end, levelDetails);
+      assert.strictEqual(actual, mockFragments[mockFragments.length - 1]);
     });
 
-    it('PDT search choosing fragment after starting/seeking to a new position (bufferEnd used)', function () {
-      levelDetails.programDateTime = PDT;// If programDateTime contains a date then PDT is used
-      let mediaSeekingTime = 17.00;
+    describe('PDT Searching during a live stream', function () {
+      before(function () {
+        levelDetails.hasProgramDateTime = true;
+      });
 
-      let foundFragment = streamController._findFragment(0, null, fragLen, mockFragments, mediaSeekingTime, end, levelDetails);
-      let resultSN = foundFragment ? foundFragment.sn : -1;
-      assert.equal(foundFragment, mockFragments[3], 'Expected sn 3, found sn segment ' + resultSN);
-    });
+      it('PDT search choosing fragment after level loaded', function () {
+        levelDetails.PTSKnown = false;
+        levelDetails.live = true;
 
-    it('PDT serch hitting empty discontinuity', function () {
-      levelDetails.programDateTime = PDT;// If programDateTime contains a date then PDT is used
-      let discontinuityPDTHit = 6.00;
+        let foundFragment = streamController._ensureFragmentAtLivePoint(levelDetails, bufferEnd, 0, end, fragPrevious, mockFragments, mockFragments.length);
+        let resultSN = foundFragment ? foundFragment.sn : -1;
+        assert.equal(foundFragment, mockFragments[2], 'Expected sn 2, found sn segment ' + resultSN);
+      });
 
-      let foundFragment = streamController._findFragment(0, null, fragLen, mockFragments, discontinuityPDTHit, end, levelDetails);
-      let resultSN = foundFragment ? foundFragment.sn : -1;
-      assert.equal(foundFragment, mockFragments[1], 'Expected sn 1, found sn segment ' + resultSN);
-    });
+      it('PDT search hitting empty discontinuity', function () {
+        let discontinuityPDTHit = 6.00;
 
-    it('finds the next fragment by SN if finding by PDT returns a frag out of tolerance', function () {
-      levelDetails.programDateTime = PDT;
-      const fragments = [mockFragments[0], mockFragments[0], mockFragments[1]];
-      const bufferEnd = fragments[1].start + fragments[1].duration;
-      const end = fragments[2].start + fragments[2].duration;
-
-      const expected = fragments[2];
-      const actual = streamController._findFragment(0, fragments[1], fragments.length, fragments, bufferEnd, end, levelDetails);
-      assert.strictEqual(expected, actual);
+        let foundFragment = streamController._ensureFragmentAtLivePoint(levelDetails, discontinuityPDTHit, 0, end, fragPrevious, mockFragments, mockFragments.length);
+        let resultSN = foundFragment ? foundFragment.sn : -1;
+        assert.equal(foundFragment, mockFragments[2], 'Expected sn 2, found sn segment ' + resultSN);
+      });
     });
   });
 
@@ -205,6 +195,76 @@ describe('StreamController tests', function () {
       fragStateStub(FragmentState.APPENDING);
       streamController._loadFragment(frag);
       assertNotLoadingState();
+    });
+  });
+
+  describe('checkBuffer', function () {
+    let sandbox;
+    beforeEach(function () {
+      sandbox = sinon.sandbox.create();
+      streamController.gapController = {
+        poll: () => {}
+      };
+      streamController.media = {
+        buffered: {
+          length: 1
+        }
+      };
+    });
+    afterEach(function () {
+      sandbox.restore();
+    });
+
+    it('should not throw when media is undefined', function () {
+      streamController.media = null;
+      streamController._checkBuffer();
+    });
+
+    it('should seek to start pos when metadata has not yet been loaded', function () {
+      const seekStub = sandbox.stub(streamController, '_seekToStartPos');
+      streamController.loadedmetadata = false;
+      streamController._checkBuffer();
+      assert(seekStub.calledOnce);
+      assert(streamController.loadedmetadata);
+    });
+
+    it('should not seek to start pos when metadata has been loaded', function () {
+      const seekStub = sandbox.stub(streamController, '_seekToStartPos');
+      streamController.loadedmetadata = true;
+      streamController._checkBuffer();
+      assert(seekStub.notCalled);
+      assert(streamController.loadedmetadata);
+    });
+
+    it('should not seek to start pos when nothing has been buffered', function () {
+      const seekStub = sandbox.stub(streamController, '_seekToStartPos');
+      streamController.media.buffered.length = 0;
+      streamController._checkBuffer();
+      assert(seekStub.notCalled);
+      assert.strictEqual(streamController.loadedmetadata, undefined);
+    });
+
+    it('should complete the immediate switch if signalled', function () {
+      const levelSwitchStub = sandbox.stub(streamController, 'immediateLevelSwitchEnd');
+      streamController.loadedmetadata = true;
+      streamController.immediateSwitch = true;
+      streamController._checkBuffer();
+      assert(levelSwitchStub.called);
+    });
+
+    describe('_seekToStartPos', function () {
+      it('should seek to startPosition when startPosition is not buffered & the media is not seeking', function () {
+        streamController.startPosition = 5;
+        streamController._seekToStartPos();
+        assert.strictEqual(5, streamController.media.currentTime);
+      });
+
+      it('should not seek to startPosition when it is buffered', function () {
+        streamController.startPosition = 5;
+        streamController.media.currentTime = 5;
+        streamController._seekToStartPos();
+        assert.strictEqual(5, streamController.media.currentTime);
+      });
     });
   });
 });
