@@ -1,5 +1,5 @@
-let assert = require('assert');
 import BinarySearch from '../../../src/utils/binary-search';
+let assert = require('assert');
 
 describe('binary search util', function () {
   describe('search helper', function () {
