@@ -1,6 +1,6 @@
-const assert = require('assert');
-
 import { BufferHelper } from '../../../src/utils/buffer-helper';
+
+const assert = require('assert');
 
 function createMockBuffer (buffered) {
   return {
